adding mongo service to the update cache function
