+++ backend/BGC.Core/Models/Settings/MongoDbSettings.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.Diagnostics.CodeAnalysis;
 
-namespace BGC.SearchApi.Models.Settings
+namespace BGC.Core.Models.Settings
 {
     [ExcludeFromCodeCoverage(Justification = "Settings model don't require testing")]
     public record MongoDbSettings
