+++ backend/BGC.SearchApi/Common/Constants.cs
@@ -5,17 +5,6 @@ namespace BGC.SearchApi.Common;
 [SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1600:ElementsMustBeDocumented", Justification = "No need to document constants, as their names should be self explanatory")]
 public static class Constants
 {
-    public static class MongoDb
-    {
-        public const string BgcDbName = "boardGames";
-        public const string BoardGamesDbCollectionName = "bgg";
-
-        public static class ConventionNames
-        {
-            public const string CamelCase = "camelCase";
-        }
-    }
-
     public static class ConfigurationKeyNames
     {
         public const string IsIntegrationTest = "IntegrationTest";
