+++ backend/BGC.SearchApi/Services/SearchService.cs
@@ -2,10 +2,10 @@ using System.Collections.Generic;
 using System.Net;
 
 using BGC.Core.Models.Domain;
+using BGC.Core.Repositories.Interfaces;
 using BGC.Core.Services.Interfaces;
 using BGC.SearchApi.Models.Dtos;
 using BGC.SearchApi.Models.Exceptions;
-using BGC.SearchApi.Repositories.Interfaces;
 using BGC.SearchApi.Services.Interface;
 using BGC.SearchApi.Services.Interfaces;
